AOSP low memory killer + Fugeswap
Change the lowmemorykiller to match android-2.6.36.y common kernel in AOSP
In addition add fudgeswap

fudgeswap acts as follows:

If set to non zero (defualt is 512k):
	Check for the amount of SWAP_FREE space avalible
	If > 0KB is avalible:
		if fudgeswap > swapfree:
			other_file += swapfree
		else:
			other_file += fugeswap

In short: we will add in fugeswap as long as its less then the free swap

Setting this to a very large positive number will indicate swap ought
to be fully used as free (and will slow the system down)

smaller numbers will allow you to put some pressure on SWAP without
slowing the system down as much.

small negitive numbers will allow the system to be faster at the same
minfree level.

default is 512 to give a very little bit of pressure to use some swap
but this can be modified at runtime via:
/sys/module/lowmemorykiller/parameters/fugeswap

-------------------------------------------------------------------

Just a slight correction to the commit details:

Where I say 512k is false .. its 512 pages, each page by default is 4k
So this is really 2MB

This is fine while next to the AOSP default 2048 min free level for interactive.. (attempts to keep 8MB free at all times)
of course that 8MB is usually mostly related to file swaps.. and eating too far into them can cause slowdown.. But on the Dream/Sapphire this amount is huge considering the other levels are much larger. (2048,3072,4096,6144,7168,8192)

I'm currently experimenting with
fudge=512 (default)
minfree=512,1024,2048,3076,4096,6144
18MB compcache (via zram)

Conflicts:

	drivers/staging/android/lowmemorykiller.c